Refactor orb.add_cog, check correctness of on_blah methods. (Now when a cog is added, the orb checks that the on_receivers correspond to real messages and that their arg spec matches the nearcast spec.)
Work out what to do with enter and backspace. (We had gone down the road of having different kinds of enumerations for special keys like enter and tab and backspace from standard characters. This causes crashes. We are now returning to unicode, but with some of the early ascii range wiped out to represent mouse click events.)

Improve mouse support in console to recognise left and right buttons. (This does not affect the current terminal, just the console. There is lots of potential follow-on work from this to console and term and gollop mode. But the upside of this commit is that the direction is now clear.)